Turned off fixed sized obs code.
